API_Rest/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2==2.9.1
coverage==5.5
PyJWT==2.1.0